Toute la partie batch est terminée. Reste à finir le front

+++ batch/src/main/java/fr/axelallain/batch/model/Copy.java
@@ -4,7 +4,7 @@ import java.util.Collection;
 
 public class Copy {
 
-    private Integer id;
+    private Long id;
 
     private boolean available;
 
@@ -15,15 +15,15 @@ public class Copy {
     public Copy() {
     }
 
-    public Copy(Integer id) {
+    public Copy(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
